db_test2/db.py
import os,certifi
from dotenv import load_dotenv
from motor.motor_asyncio import AsyncIOMotorClient
load_dotenv()

MONGO_URI = os.getenv("MONGO_URI")

client = AsyncIOMotorClient(
    MONGO_URI,
    tls= True,
    tlsCAFile=certifi.where()
)

db = client.fastapi_db
users_collection = db.users