db_test2/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
dnspython==2.8.0
fastapi==0.128.0
h11==0.16.0
idna==3.11
motor==3.7.1
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dotenv==1.2.1
starlette==0.50.0
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
